# Conflicts:
#	Assets/Client/Scripts/Client/Contents/Minion/Attack/BattleAbility.cs
#	Assets/Client/Scripts/Client/Contents/MinionDataBaseIngame.cs
#	Assets/Server/Scripts/SDKInitializer.cs

+++ Assets/Client/Scripts/Client/Contents/PlayerData.cs
@@ -12,7 +12,7 @@ public class PlayerData : IMinionDeployable
     private string playerSessionID;
     private ulong clientID;
 
-    public int currentGem;
+    public float currentGem;
 
     public IReadOnlyList<MinionData> MinionDeck => minionDeck;
